Finish moving `lib` to `src`
